GET/DROP commands implemented
The Container, Portable, Sentience, and Parser classes have all been created or modified to allow picking up Portable things and dropping them.

Note that Tiles currently allow only ONE item at a time! A formal item container will be implemented when the GUI allows for the retrieval of this information.

On branch develop

Changes to be committed:
	modified:   src/action.cpp
	modified:   src/action.hpp
	modified:   src/actor.cpp
	modified:   src/actor.hpp
	modified:   src/container.cpp
	modified:   src/container.hpp
	modified:   src/context.cpp
	modified:   src/context.hpp
	modified:   src/engine.cpp
	modified:   src/item.cpp
	modified:   src/map.cpp
	modified:   src/map.hpp
	modified:   src/parser.cpp
	modified:   src/portable.cpp
	modified:   src/sentience.cpp